Add files via upload
I've rewritten the code for the TOS_S@_DESeq2.R script to address two issues.

The way that the DESeq() comparison was made originally was actually comparing all doses of one of the drugs to the control to evaluate for differences. This has been changed by how the group variable is defined. The new line for this is qc_hash$group <- paste0(qc_hash$Treatment, "_", round(qc_hash$numeric_dose,0)). NOTE: This causes the number of DEGs to be different from the examples.
The lfcShrink() changes made to DESeq2 where we really want L2FC values from lfcShrink(), but pvalues from results(). I've written a function DESeq2_sContrast() that does the following:
Create a MA Plot showing the Differentially expressed genes for the comparions
Use the lfcShrink() function to get the l2fc values
Use the results() function to get pvalues, adj.pvalues etc..
Combined into a dataset for all genes tested for that comparison and write a .csv
Return that dataset to an object in memory